+++ boot.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
 $LOAD_PATH << File.expand_path("lib", __dir__)
-require "historical_currency"
+require "currency_patches"